backend/requirements.txt
Flask==3.1.2
Flask-JWT-Extended==4.7.1
Flask-Cors==6.0.2
python-dotenv==1.2.1
Flask-SQLAlchemy==3.1.1
mysql-connector-python==9.5.0
google-auth==2.47.0
google-auth-oauthlib==1.2.4
requests==2.32.5
openai==2.15.0
deepl==1.27.0
PyPDF2==3.0.1
python-docx==1.2.0
openpyxl==3.1.5
fpdf==1.7.2
Werkzeug==3.1.5